--- challenge/challenge/models/model_test/__init__.py
@@ -1,1 +1,0 @@
-from .model import *
