+++ experiments/config.yml
@@ -1,10 +1,10 @@
-name: baseline
+name: test_model
 save_dir: saved/
 seed: 1234
 target_devices: [0]
 
 arch:
-  type: Baseline
+  type: TestModel
   args:
     in_features: 1280
